feat: move mutate handlers into utils
